Add files via upload
This file contains a descriptive code on how to generate additive white gaussian noise. The purpose is to help one understand the concepts and you can code up an additive gaussian white noise. a simple analysis was performed to calculate the Fourier transform of the additive gaussian white noise and display the results.